feat: migrate pack and resize
